requirements.txt
aniso8601          8.0.0
bcrypt             3.2.0
boto3              1.21.45
botocore           1.24.45
certifi            2021.10.8
cffi               1.15.0
charset-normalizer 2.0.12
click              7.1.2
Flask              1.1.2
Flask-Cors         3.0.10
Flask-RESTful      0.3.8
Flask-SQLAlchemy   2.4.3
idna               3.3
importlib-metadata 4.11.3
itsdangerous       1.1.0
Jinja2             2.11.2
jmespath           1.0.0
MarkupSafe         1.1.1
pip                22.0.4
psycopg2           2.9.3
pycparser          2.21
PyJWT              2.3.0
python-dateutil    2.8.2
pytz               2020.1
requests           2.27.1
s3transfer         0.5.2
setuptools         61.1.1
six                1.15.0
SQLAlchemy         1.3.18
urllib3            1.26.9
Werkzeug           1.0.1
wheel              0.37.1
zipp               3.8.0
